Remove test_expired_file_delete to own file
